requirements.txt
psycopg2-binary
pandas
flask
requests
beautifulsoup4